+++ c_32bit/mtalloc/alloc.h
@@ -6,10 +6,10 @@
 
 #include <stdint.h>
 
-void alloc_free(uint16_t const block_addr);
+void alloc_free(uint32_t const block_addr);
 
-uint16_t alloc_alloc(uint16_t const wanted_len);
+uint32_t alloc_alloc(uint32_t const wanted_len);
 
-void alloc_init(uint8_t * const mem, uint16_t const mem_len);
+void alloc_init(uint8_t * const mem, uint32_t const mem_len);
 
 #endif //MT_ALLOC
